Finished NPM Integration and also implemented MongoDB interaction.
